AP-14577: Clean up unused method parameters.
The convenience methods joinOutputCombined and joinOutputSplit can be provided by the base class JoinImplementation.
Move the test for disjuntive joining into BlackHashJoin#join to make sure the join specification is accounted for (disjunctive join would only be tested for in the convenience methods joinOutputCombined and joinOutputSplit).

AP-14577 (Join with Match Any)